+++ server/api/tasks_api.py
@@ -2,48 +2,7 @@ from flask_restful import Resource, current_app, request
 from server.services.task_service import TaskService, TaskServiceError
 
 
-class TaskAPI(Resource):
 
-    def get(self, project_id, task_id):
-        """
-        Get task details
-        ---
-        tags:
-            - tasks
-        produces:
-            - application/json
-        parameters:
-            - name: project_id
-              in: path
-              description: The ID of the project the task is associated with
-              required: true
-              type: integer
-              default: 1
-            - name: task_id
-              in: path
-              description: The unique task ID
-              required: true
-              type: integer
-              default: 1
-        responses:
-            200:
-                description: Task found
-            404:
-                description: Task not found
-            500:
-                description: Internal Server Error
-        """
-        try:
-            task = TaskService().get_task_as_dto(task_id, project_id)
-
-            if task is None:
-                return {"Error": "Task Not Found"}, 404
-
-            return task, 200
-        except Exception as e:
-            error_msg = f'Task GET API - unhandled error: {str(e)}'
-            current_app.logger.critical(error_msg)
-            return {"Error": error_msg}, 500
 
 
 class UnlockTaskAPI(Resource):
